split out platforms
