+++ engine/platform/platform.h
@@ -4,33 +4,6 @@
 
 namespace egkr
 {
-
-	enum class platform_type
-	{
-		windows,
-		linux,
-		macos,
-		android
-	};
-
-	inline std::string_view platform_type_to_string(platform_type type)
-	{
-		switch (type)
-		{
-		case egkr::platform_type::windows:
-			return "windows";
-		case egkr::platform_type::linux:
-			return "linux";
-		case egkr::platform_type::macos:
-			return "macos";
-		case egkr::platform_type::android:
-			return "android";
-		default:
-			LOG_ERROR("Unknown platform type");
-			return "";
-		}
-	}
-
 	class platform : public std::enable_shared_from_this<platform>
 	{
 	public:
@@ -40,10 +13,10 @@ namespace egkr
 			uint32_t start_y{};
 			uint32_t width_{};
 			uint32_t height_{};
-			std::string name{};
+			std::string name;
 		};
 		using shared_ptr = std::shared_ptr<platform>;
-		static shared_ptr create(platform_type type);
+		static shared_ptr create();
 
 		virtual ~platform() = default;
